moves Counter and Chronometer to utils
