+++ apps/stopwatch/main.py
@@ -1,63 +1,8 @@
 # coding=utf-8
-from time import time
 
 from apps.zero_app import ZeroApp
 from ui import Refresher
-
-
-class Chronometer(object):
-    def __init__(self):
-        self.__active = False
-        self.__last_call = time()
-
-    def tick(self):
-        """
-        :rtype: int
-        :return: the time elapsed since the previous tick
-        """
-        now = time()
-        elapsed = now - self.__last_call
-        self.__last_call = now
-        return elapsed
-
-
-class Counter(object):
-    def __init__(self):
-        self.__active = False
-        self.__cron = Chronometer()
-        self.__elapsed = 0
-
-    @property
-    def active(self):
-        return self.__active
-
-    @property
-    def elapsed(self):
-        return self.__elapsed
-
-    def update(self):
-        if not self.__active:
-            return
-        self.__elapsed += self.__cron.tick()
-
-    def stop(self):
-        self.__cron.tick()
-        self.__elapsed = 0
-        self.__active = False
-
-    def pause(self):
-        self.__active = False
-
-    def resume(self):
-        self.__cron.tick()
-        self.__active = True
-
-    def start(self):
-        self.stop()
-        self.resume()
-
-    def toggle(self):
-        self.pause() if self.active else self.resume()
+from ui.utils import Counter
 
 
 class StopwatchApp(ZeroApp):
